feat(firmware): p.o.c to sent an e-stop command to an ESP over HTTP
Uses the ESP in Access Point mode and hosts a tiny web server with a
single plain HTML form. Through this page you can send commands as URL
query parameters, which will be handled by the ESP to control a CO-SPDT
relay for switching the power of a circuit.

Documentation will follow.